% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{as_dummy_df}
\alias{as_dummy_df}
\title{Title}
\usage{
as_dummy_df(data_inp, col_cell = "cell_names")
}
\arguments{
\item{data_inp}{Cell-coordinates matrix}

\item{col_cell}{Column name of cell type, cell type names must be syntactically valid}
}
\value{
dummy data frame with cells on columns
}
\description{
Title
}
\examples{
cell_dummy_df <- as_dummy_df(celltrek_df, col_cell='Cell')
}
