% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{build_delaunayn}
\alias{build_delaunayn}
\title{Title}
\usage{
build_delaunayn(coord_df, return_name = T, dist_cutoff = NULL)
}
\arguments{
\item{coord_df}{Coordinates df}

\item{return_name}{Return edge list with names?}

\item{dist_cutoff}{Remove edges with distance >= dist_cutoff}
}
\value{
Delaunay triangulation network edge list
}
\description{
Title
}
\examples{
test <- build_delaunayn(coord_df, return_name=T)
}
