% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{cc_wrapper}
\alias{cc_wrapper}
\title{A wrapper function of ConensusClusterPlus with some default parameters}
\usage{
cc_wrapper(
  d,
  maxK = 8,
  reps = 20,
  distance = "spearman",
  verbose = F,
  plot = "png",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
A wrapper function of ConensusClusterPlus with some default parameters
}
\examples{
cc_wrapper(d, maxK=8, reps=20, distance='spearman', verbose=T)
}
