% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek.R
\name{celltrek}
\alias{celltrek}
\title{Title}
\usage{
celltrek(
  st_sc_int,
  int_assay = "traint",
  sc_data = NULL,
  sc_assay = "RNA",
  reduction = "pca",
  intp = T,
  intp_pnt = 10000,
  intp_lin = F,
  nPCs = 30,
  ntree = 1000,
  dist_thresh = 0.4,
  top_spot = 10,
  spot_n = 10,
  repel_r = 5,
  repel_iter = 10,
  keep_model = F,
  ...
)
}
\arguments{
\item{st_sc_int}{Seurat traint object}

\item{int_assay}{Integration assay ('traint')}

\item{sc_data}{SC data, optional}

\item{sc_assay}{SC assay}

\item{reduction}{Dimension reduction method, usually 'pca'}

\item{intp}{If True, do interpolation}

\item{intp_pnt}{Number of interpolation points}

\item{intp_lin}{If Ture, do linear interpolation}

\item{nPCs}{Number of PCs}

\item{ntree}{Number of Trees}

\item{dist_thresh}{Distance threshold}

\item{top_spot}{Maximum number of spots that one cell can be charted}

\item{spot_n}{Maximum number of cells that one spot can contain}

\item{repel_r}{Repelling radius}

\item{repel_iter}{Repelling iterations}

\item{keep_model}{If TRUE, return the trained random forest model}

\item{...}{}
}
\value{
Seurat object
}
\description{
Title
}
\examples{
celltrek_res <- celltrek(st_sc_int, int_assay='traint', sc_data=NULL, sc_assay='RNA', reduction='pca', intp=T, intp_pnt=10000, intp_lin=F, nPCs=30, ntree=1000, dist_thresh=.4, top_spot=10, spot_n=10, r=NULL, keep_model=F, ...)
}
