% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek.R
\name{celltrek_chart}
\alias{celltrek_chart}
\title{Title}
\usage{
celltrek_chart(
  dist_mat,
  coord_df,
  dist_cut = 500,
  top_spot = 10,
  spot_n = 10,
  repel_r = 5,
  repel_iter = 10
)
}
\arguments{
\item{dist_mat}{Distance matrix of sc-st (sc in rows and st in columns)}

\item{coord_df}{Coordinates data frame of st (must contain coord_x, coord_y columns, barcode rownames)}

\item{dist_cut}{Distance cutoff}

\item{top_spot}{Maximum number of spots that one cell can be charted}

\item{spot_n}{Maximum number of cells that one spot can contain}

\item{repel_r}{Repelling radius}

\item{repel_iter}{Repelling iterations}
}
\value{
SC coordinates
}
\description{
Title
}
\examples{
chart_test <- function (dist_mat, coord_df, dist_cut=500, top_spot=10, spot_n=10, r=50)
}
