% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek.R
\name{celltrek_dist}
\alias{celltrek_dist}
\title{Title Calculate the RF-distance between sc and st}
\usage{
celltrek_dist(
  st_sc_int,
  int_assay = "traint",
  reduction = "pca",
  intp = T,
  intp_pnt = 10000,
  intp_lin = F,
  nPCs = 30,
  ntree = 1000,
  keep_model = T
)
}
\arguments{
\item{st_sc_int}{Seurat traint object}

\item{int_assay}{Name of integration assay}

\item{reduction}{Dimension reduction method used, usually pca}

\item{intp}{If TRUE, do interpolation}

\item{intp_pnt}{Interpolation point number}

\item{intp_lin}{If TRUE, use linear interpolation}

\item{nPCs}{Number of PCs used for CellTrek}

\item{ntree}{Number of trees in random forest}

\item{keep_model}{If TRUE, return the trained random forest model}
}
\value{
A list of 1. celltrek_distance matrix; 2. trained random forest model (optional)
}
\description{
Title Calculate the RF-distance between sc and st
}
\examples{
dist_test <- celltrek_dist(st_sc_int=st_sc_int, int_assay='traint', reduction='pca', intp = T, intp_pnt=10000, intp_lin=F, nPCs=30, ntree=1000, keep_model=T)
}
