% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek.R
\name{celltrek_repel}
\alias{celltrek_repel}
\title{Title Mannually repelling celltrek cells}
\usage{
celltrek_repel(celltrek_inp, repel_r = 5, repel_iter = 10)
}
\arguments{
\item{celltrek_inp}{CellTrek(Seurat) object}

\item{repel_r}{Repelling radius}

\item{repel_iter}{Repelling iterations}
}
\value{
CellTrek(Seurat) object
}
\description{
Title Mannually repelling celltrek cells
}
