% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek_vis.R
\name{celltrek_vis}
\alias{celltrek_vis}
\title{Visualization of CellTrek results}
\usage{
celltrek_vis(celltrek_df, img, scale_fac)
}
\arguments{
\item{celltrek_df}{CellTrek output meta data, must contain coord_x and coord_y}

\item{img}{Image data}

\item{scale_fac}{Scale factor}
}
\value{
Shiny GUI
}
\description{
Visualization of CellTrek results
}
\examples{
celltrek_vis(test_celltrek@meta.data, test_celltrek@images[[1]]@image, scale_fac=test_celltrek@images[[1]]@scale.factors$lowres)
}
