% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{cor_remove}
\alias{cor_remove}
\title{Remove low correlation genes until reaching threshold}
\usage{
cor_remove(cor_mat, ave_cor_cut = 0.5, min_n = 5, max_n = 100, na_diag = F)
}
\arguments{
\item{na_diag}{}
}
\value{

}
\description{
Remove low correlation genes until reaching threshold
}
\examples{
cor_remove(cor_mat, ave_cor_cut = 0.5, min_n=10, max_n=100, na_diag=F)
}
