% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{DT_boot_mst}
\alias{DT_boot_mst}
\title{Title}
\usage{
DT_boot_mst(
  meta_df,
  coord_df,
  col_cell = "cell_names",
  boot_n = 100,
  prop = 0.8,
  replace = T,
  dist_cutoff = NULL
)
}
\arguments{
\item{meta_df}{Meta data, must contain cell type column}

\item{coord_df}{Coordinates data, must be the same order of meta_df}

\item{col_cell}{Column name of cell type, cell type names must be syntactically valid}

\item{boot_n}{Number of bootstrapping iterations}

\item{prop}{Subsample proportion}

\item{replace}{should sampling be with replacement?}

\item{dist_cutoff}{Remove edges with distance >= dist_cutoff}
}
\value{
A list of 1. Bootstrap logOR distance; 2. MST consensus matrix
}
\description{
Title
}
\examples{
test <- DT_boot_mst(meta_df, coord_df, col_cell='cell_names', boot_n=100, prop=.8)
}
