% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{edge_odds}
\alias{edge_odds}
\title{Title}
\usage{
edge_odds(el_inp, meta_data, col_cell = "cell_names")
}
\arguments{
\item{el_inp}{Edge list of Delaunay triangulation network}

\item{meta_data}{Meta data, must contain column of cell type}

\item{col_cell}{Column name of cell type, cell type names must be syntactically valid}
}
\value{
A list of 1.Network-based cell co-occurrence; 2.Cell co-occurrence log odds ratio
}
\description{
Title
}
\examples{
test <- edge_odds(el_inp, meta_data, col_cell='cell_names')
}
