% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCorMarkers.R
\name{FindCorMarkers}
\alias{FindCorMarkers}
\title{Find highly correlated markers to target features}
\usage{
FindCorMarkers(srt_inp, assay = "RNA", features, method = "spearman")
}
\arguments{
\item{srt_inp}{Seurat input}

\item{features}{Features that genes correlated with}

\item{method}{Correlation method}
}
\value{

}
\description{
Find highly correlated markers to target features
}
\examples{
CorMarkers <- FindCorMarkers(srt_inp=test_seurat, assay='RNA', features=c('k_dist', 'sig_score'), method='spearman')
}
