% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek_vis.R
\name{img_gs}
\alias{img_gs}
\title{Convert RGB image to gray scale image}
\usage{
img_gs(img)
}
\arguments{
\item{img}{Image data}
}
\value{
A gray scale image
}
\description{
Convert RGB image to gray scale image
}
\examples{
img_gs <- img_gs(img_raw)
}
