% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{KL_}
\alias{KL_}
\title{Title}
\usage{
KL_(X, eps = 1e-20)
}
\arguments{
\item{X}{Density matrix}

\item{eps}{Small value added}
}
\value{
KL-divergence matrix
}
\description{
Title
}
\examples{
KL_res <- KL_(X=X, eps=1e-20)
}
