% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{KL_boot_mst}
\alias{KL_boot_mst}
\title{Title}
\usage{
KL_boot_mst(
  dummy_df,
  coord_df,
  min_num = 15,
  Xlim = range(coord_df[, 1]),
  Ylim = range(coord_df[, 2]),
  boot_n = 100,
  prop = 0.8,
  replace = T,
  h = 150,
  n = 25,
  tot_norm = T,
  eps = 1e-20
)
}
\arguments{
\item{dummy_df}{Cell type dummy df}

\item{coord_df}{Coordinates df}

\item{min_num}{For cells number < min_num, add 1 till min_num}

\item{Xlim}{The limits of X-axis}

\item{Ylim}{The limits of Y-axis}

\item{boot_n}{Number of bootstrapping iterations}

\item{prop}{Subsample proportion}

\item{replace}{should sampling be with replacement?}

\item{h}{Bandwidths for x and y directions, more details in kde2d function in MASS package}

\item{n}{Number of grid points in each directions, more details in kde2d function in MASS package}

\item{tot_norm}{Normalization by total?}

\item{eps}{Small value when calculate KL-divergence}
}
\value{
A list of 1. Bootstrap KL-divergence; 2. MST consensus matrix
}
\description{
Title
}
\examples{
boot_mst_res <- KL_boot_mst(dummy_df=cell_type_dummy, coord_df=range(coord_df$X), Xlim=range(coord_df$Y), boot_n=100, prop=0.8, h=150, n=25, tot_norm=T, eps=1e-20)
}
