% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{rbfk}
\alias{rbfk}
\title{Radial basis function kernel}
\usage{
rbfk(dis_mat, sigm, zero_diag = T)
}
\arguments{
\item{dis_mat}{Distance matrix}

\item{sigm}{Width of rbfk}

\item{zero_diag}{}
}
\value{

}
\description{
Radial basis function kernel
}
\examples{
rbfk(dis_mat, sigm, zero_diag=F)
}
