% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdist.R
\name{run_kdist}
\alias{run_kdist}
\title{Title}
\usage{
run_kdist(
  schart_inp,
  grp_col = "cell_type",
  ref = NULL,
  ref_type = "all",
  que = NULL,
  k = 10,
  new_name = "kdist",
  keep_nn = F
)
}
\arguments{
\item{schart_inp}{SChart seurat input}

\item{grp_col}{Column name in meta data for reference and query groups}

\item{ref}{Reference groups}

\item{ref_type}{'all' or 'each'}

\item{que}{Query groups}

\item{k}{The number of nearest neighbors}

\item{new_name}{New name for kdist}

\item{keep_nn}{Keep Nearest Neighboor id matrix?}
}
\value{
SChart seurat output
}
\description{
Title
}
\examples{
schart_test <- run_kdist(schart_inp=schart_inp, grp_col='cell_type', ref=c('A', 'B'), ref_type='each', que=unique(test_df$cell_names), k=10)
}
