% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schart.R
\name{schart_raw}
\alias{schart_raw}
\title{Title}
\usage{
schart_raw(
  st_sc_int,
  int_assay = "traint",
  sc_data = NULL,
  sc_assay = "RNA",
  edgelist_return = T,
  reduction = "pca",
  nPCs = 30,
  ntree = 1000,
  dist_thresh = 0.4,
  top_spots = 5,
  spot_n = 20,
  keep_model = F,
  ...
)
}
\arguments{
\item{st_sc_int}{traint object}

\item{int_assay}{integration assay}

\item{sc_data}{SC data, optional}

\item{sc_assay}{SC assay}

\item{edgelist_return}{If TRUE, return ST-SC sparse graph (as edgelist)}

\item{reduction}{Dimension reduction method used, usually pca}

\item{nPCs}{Number of principal components used for SChart}

\item{ntree}{Number of trees in random forest}

\item{top_spots}{Maximum number of spots that one cell can be charted}

\item{spot_n}{Maximum number of cells that one spot can contain}

\item{keep_model}{If TRUE, return the trained random forest model}

\item{...}{}

\item{dist_th}{Random forest distance threshold}
}
\value{
Seurat object
}
\description{
Title
}
\examples{
SChart_res <- schart_raw(st_sc_int=st_sc_traint, int_assay='traint', sc_data=NULL, sc_assay='RNA', edgelist_return=T, reduction='pca', nPCs=30, ntree=1000, dist_thresh=.4, top_spots=5, spot_n=20, keep_model=T)
}
