% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schart_vis.R
\name{schart_vis}
\alias{schart_vis}
\title{Title}
\usage{
schart_vis(schart_df, img, scale_fac)
}
\arguments{
\item{schart_df}{SChart output meta data, must contain coord_x and coord_y}

\item{img}{Image data}

\item{scale_fac}{Scale factor}
}
\value{
Shiny GUI
}
\description{
Title
}
\examples{
schart_vis(test_schart@meta.data, test_schart@images[[1]]@image, scale_fac=test_schart@images[[1]]@scale.factors$lowres)
}
