% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc_vis.R
\name{scoloc_vis}
\alias{scoloc_vis}
\title{Visualization of SColoc results}
\usage{
scoloc_vis(adj_mat, meta_data = NULL, directed = F)
}
\arguments{
\item{adj_mat}{Adjacent matrix}

\item{meta_data}{Optional, must contain id column match with adj_mat col/rownames}

\item{directed}{Generate a directed graph? (Default is false)}
}
\value{
shiny app
}
\description{
Visualization of SColoc results
}
\examples{
scoloc_vis(scoloc_test)
}
