% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{sp_grid_kern_bin}
\alias{sp_grid_kern_bin}
\title{Title}
\usage{
sp_grid_kern_bin(
  data,
  coord,
  min_num = 15,
  h,
  n = 25,
  tot_norm = TRUE,
  Xlim = range(coord[, 1]),
  Ylim = range(coord[, 2])
)
}
\arguments{
\item{data}{Cell type dummy table}

\item{coord}{Coordinates data with X and Y columns}

\item{min_num}{For cells number < min_num, add 1 till min_num}

\item{h}{Bandwidths for x and y directions, more details in kde2d function in MASS package}

\item{n}{Number of grid points in each directions, more details in kde2d function in MASS package}

\item{tot_norm}{Normalization by total?}

\item{Xlim}{The limits of X-axis}

\item{Ylim}{The limits of Y-axis}
}
\value{
Kenerl density on grid
}
\description{
Title
}
\examples{
kern_res <- sp_grid_kern_bin(data=cell_dummy_df, coord=coord_df, h=150, n=25, tot_norm=TRUE, Xlim=range(coord_df$X), Ylim=range(coor_df$Y))
}
