% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{wcor}
\alias{wcor}
\title{Title}
\usage{
wcor(X, W, method = c("pearson", "spearman")[1], na_zero = T)
}
\arguments{
\item{X}{Expression matrix, n X p}

\item{W}{Weight matrix, n X n}

\item{method}{Correlation method, pearson or spearman}

\item{na_zero}{Na to zero}
}
\value{
Weighted correlation matrix, p X p
}
\description{
Title
}
\examples{
wcor(X=expr_test, W=rbfk_out, method='spearman')
}
