% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{wgcna_gene_k}
\alias{wgcna_gene_k}
\title{Title}
\usage{
wgcna_gene_k(
  wgcna_res,
  cor_mat,
  avg_cor_min = 0.5,
  min_gen = 10,
  max_gen = 100
)
}
\arguments{
\item{max_gen}{}
}
\value{

}
\description{
Title
}
\examples{
wgcna_gene_k(wgcna_res, cor_mat, avg_cor_min=.5, min_gen=10, max_gen=100)
}
