% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdist.R
\name{kdist}
\alias{kdist}
\title{Title}
\usage{
kdist(
  inp_df,
  ref = NULL,
  ref_type = "all",
  que = NULL,
  k = 10,
  new_name = "kdist",
  keep_nn = F
)
}
\arguments{
\item{inp_df}{inp_df must contain cell_names, 'coord_x', 'coord_y' columns}

\item{ref}{Reference groups}

\item{ref_type}{'all' or 'each'}

\item{que}{Query groups}

\item{k}{The number of nearest neighbors}

\item{new_name}{New name for kdist}

\item{keep_nn}{Keep Nearest Neighboor id matrix?}
}
\value{
A list of 1. kdist data frame and 2. a list of knn id matrix
}
\description{
Title
}
\examples{
kdist_out <- kdist(inp_df=test_df, ref=c('A', 'B', 'C'), ref_type='each', que=unique(test_df$cell_names), k=10, keep_nn=T)
}
