% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoloc.R
\name{scoloc}
\alias{scoloc}
\title{SColoc module}
\usage{
scoloc(
  celltrek_inp,
  col_cell = "cell_names",
  use_method = c("KL", "DT", "KD")[2],
  h = celltrek_inp@images[[1]]@scale.factors$spot_dis,
  n = 25,
  boot_n = 20,
  ...
)
}
\arguments{
\item{celltrek_inp}{CellTrek input}

\item{col_cell}{Column name of cell type, cell type names must be syntactically valid}

\item{use_method}{Use density-based Kullback Leibler divergence or Delaunay Triangulation network}

\item{h}{Bandwidths for x and y directions, for KL, more details in kde2d function in MASS package}

\item{n}{Number of grid points in each directions, for KL, more details in kde2d function in MASS package}

\item{...}{See in boot_mst}
}
\value{
A list of 1.Bootstrap distance; 2.MST consensus matrix
}
\description{
SColoc module
}
\examples{
cell_scoloc <- scoloc(celltrek_inp, col_cell='cell_names', use_method='KL', h=140, n=25)
}
