% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrek.R
\name{traint}
\alias{traint}
\title{ST and SC data integration using Seurat transfer anchors}
\usage{
traint(
  st_data,
  sc_data,
  st_assay = "Spatial",
  sc_assay = "scint",
  norm = "LogNormalize",
  nfeatures = 2000,
  cell_names = "cell_names",
  coord_xy = c("imagerow", "imagecol"),
  gene_kept = NULL,
  ...
)
}
\arguments{
\item{st_data}{Seurat ST data object}

\item{sc_data}{Seurat SC data object}

\item{st_assay}{ST assay}

\item{sc_assay}{SC assay}

\item{norm}{normalization method: LogNormalize/SCTransform}

\item{nfeatures}{number of features for integration}

\item{cell_names}{cell cluster/type column name in SC meta data}

\item{coord_xy}{coordinates column names in ST images slot}

\item{gene_kept}{selected genes to be kept during integration}
}
\value{
Seurat object of SC-ST co-embeddings
}
\description{
ST and SC data integration using Seurat transfer anchors
}
\examples{
st_sc_traint <- traint(st_data=brain_st, sc_data=brain_sc, st_assay='Spatial', sc_assay='scint', nfeatures=2000, cell_names='cell_names', coord_xy=c('imagerow', 'imagecol'), gene_kept=NULL)
}
