% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoexp.R
\name{wgcna_wrapper}
\alias{wgcna_wrapper}
\title{Title}
\usage{
wgcna_wrapper(sim_mat, powerVector = c(1:20), minClusterSize = 50, ...)
}
\arguments{
\item{sim_mat}{similarity matrix}

\item{powerVector}{power vector}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
wgcna_wrapper(sim_mat, powerVector=c(1:20), minClusterSize=50)
}
